allow changing of request body content
