+++ src/app/views/common/monaco/Monaco.tsx
@@ -1,10 +1,11 @@
 import React from 'react';
-import MonacoEditor from 'react-monaco-editor';
+import MonacoEditor, { ChangeHandler } from 'react-monaco-editor';
 
 import './monaco.scss';
 
 interface IMonaco {
   body: object|undefined;
+  onChange?: ChangeHandler | undefined;
 }
 
 function editorDidMount(editor: any) {
@@ -24,7 +25,7 @@ function formatDocument(editor: any) {
 }
 
 export function Monaco(props: IMonaco) {
-  const { body } = props;
+  const { body, onChange } = props;
 
   return (
     <div className='monaco-editor'>
@@ -35,6 +36,7 @@ export function Monaco(props: IMonaco) {
         language='json'
         options={{ lineNumbers: 'off', minimap: { enabled: false } }}
         editorDidMount={editorDidMount}
+        onChange={onChange}
       />
     </div>
   );
