+++ src/app/views/query-runner/request/Request.tsx
@@ -1,91 +1,44 @@
 import { Pivot, PivotItem, PivotLinkSize } from 'office-ui-fabric-react';
 import React, { Component } from 'react';
+import { Monaco } from '../../common/monaco/Monaco';
 import './request.scss';
-import { RequestBodyControl } from './RequestBody';
 import { RequestHeadersControl } from './RequestHeaders';
 
-export class Request extends Component<any, any> {
-    public state = {
-        headers: [
-            { name: '', value: '' },
-        ],
-        headerName: '',
-        headerValue: '',
-    };
+interface IRequestComponent {
+    handleOnEditorChange: Function;
+    handleOnHeaderNameChange: Function;
+    handleOnHeaderDelete: Function;
+    handleOnHeaderValueChange: Function;
+    handleOnHeaderValueBlur: Function;
+  }
 
-    private handleOnHeaderNameChange = (name?: any) => {
-        if (name) {
-            this.setState({
-                headerName: name,
-            });
-        }
-    };
+export const Request = ({
+    handleOnEditorChange,
+    handleOnHeaderNameChange,
+    handleOnHeaderDelete,
+    handleOnHeaderValueChange,
+    handleOnHeaderValueBlur,
+  }: IRequestComponent) => {
 
-    public editorChange(value: any) {
-        return;
-    }
-
-    private handleOnHeaderValueChange = (value?: any) => {
-        if (value) {
-            this.setState({
-                headerValue: value,
-            });
-        }
-    };
-
-    private handleOnHeaderDelete = (headerIndex: any) => {
-        const { headers } = this.state;
-        const headersToDelete = [...headers];
-        headersToDelete.splice(headerIndex, 1);
-        this.setState({
-            headers: headersToDelete,
-        });
-        const listOfHeaders = headers;
-        if (listOfHeaders.length === 0) {
-            listOfHeaders.push({ name: '', value: '' });
-        }
-        this.setState({
-            headers: listOfHeaders,
-        });
-    };
-
-    private handleOnHeaderValueBlur = () => {
-        if (this.state.headerName !== '') {
-            const { headerName, headerValue, headers } = this.state;
-            const header = { name: headerName, value: headerValue };
-            const newHeaders = [header, ...headers];
-            this.setState({
-                headers: newHeaders,
-                headerName: '',
-                headerValue: '',
-            });
-        }
-    };
-
-    public getLastHeader() {
-        const headersLength = this.state.headers.length;
-        return this.state.headers[headersLength - 1];
-    }
-
-    public render() {
-        const { headers } = this.state;
-        return (
-            <div className='request-editors'>
-                <Pivot>
-                    <PivotItem headerText='Request Body'>
-                        <RequestBodyControl />
-                    </PivotItem>
-                    <PivotItem headerText='Request Headers'>
-                        <RequestHeadersControl
-                            handleOnHeaderDelete={this.handleOnHeaderDelete}
-                            handleOnHeaderNameChange={this.handleOnHeaderNameChange}
-                            handleOnHeaderValueChange={this.handleOnHeaderValueChange}
-                            handleOnHeaderValueBlur={this.handleOnHeaderValueBlur}
-                            headers={headers}
-                        />
-                    </PivotItem>
-                </Pivot>
-            </div>
-        );
-    }
-}
+    return (
+      <div className='request-editors'>
+        <Pivot>
+          <PivotItem headerText='Request Body'>
+            <Monaco
+                    body={undefined}
+                    onChange={(value, event) => handleOnEditorChange(value)} />
+                />
+          </PivotItem>
+          <PivotItem headerText='Request Headers'>
+            <RequestHeadersControl
+              handleOnHeaderDelete={(event: any, header: any) => handleOnHeaderDelete(header)}
+              handleOnHeaderNameChange={(event: any, name: any) => handleOnHeaderNameChange(name)}
+              handleOnHeaderValueChange={(event: any, value: any) => handleOnHeaderValueChange(value)}
+              handleOnHeaderValueBlur={(event: any, header: any) => handleOnHeaderValueBlur(header)}
+              headers={[{ name: '', value: '' }]}
+            />
+          </PivotItem>
+        </Pivot>
+      </div>
+    );
+  };
